Improve JS build system
This is a major change and makes my life a ton easier by using esbuild
as my ES6 module build system, then run Bable and everything else on top
using Gulp. Not perfectly ideal, but helps reduce my dependence on Gulp
plugins that are no longer maintained.
I eventually want to move off gulp completely, but there are a lot of
moving parts to keep track of so this will have to do for now.
Very exciting none the less!